This was the new draft. Updated for Halloween 2015.
Our family is living on Wolfe Island for the winter of 2015-2016. It is our first North American winter, and being that it is our first time trick-or-treating tonight I thought it fitting to update the story and get it done. This update was made to complete a long-forgotten story, one I put on the shelf a while back.